pyproject.toml
[tool.poetry]
authors = ["Eric Abruzzese <eric.abruzzese@gmail.com>"]
description = "A reusable Django app for managing database-backed forms."
license = "MIT"
name = "django-flexible-forms"
packages = [
  {include = "flexible_forms"},
]
version = "0.1.0"

[tool.poetry.dependencies]
django = ">=2.2"
python = "^3.8"
swapper = "^1.1.2"

[tool.poetry.dev-dependencies]
Pillow = "^7.2.0"
autopep8 = "^1.5.4"
django-stubs = "^1.5.0"
factory_boy = "^3.0.1"
hypothesis = {extras = ["django"], version = "^5.26.0"}
isort = "^5.4.2"
mypy = ">=0.770"
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
pytest-django = "^3.9.0"
pytest-randomly = "^3.4.1"
pytest-timeout = "^1.4.2"

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
